refactor(tests): extract ArgumentEchoLocalizer into shared test support module
Replaced inline ClapStubLocalizer and ClapAwareLocalizer test implementations with a new
ArgumentEchoLocalizer defined in a shared `tests/support/localizers.rs` file. Updated references
in tests and behavior steps to use this centralized localizer to reduce duplication and improve
consistency across localization tests.
